Modify the OR-clause handling so that it can safely deal with OR-clause with 17 or more terms.  Fix for ticket [10fb063b1179be53ea0b53bb].
FossilOrigin-Name: a67b5312f63909737c052fe58ab5772d45530d49

+++ src/sqliteInt.h
@@ -3019,7 +3019,7 @@ int sqlite3BitvecBuiltinTest(int,int*);
 RowSet *sqlite3RowSetInit(sqlite3*, void*, unsigned int);
 void sqlite3RowSetClear(RowSet*);
 void sqlite3RowSetInsert(RowSet*, i64);
-int sqlite3RowSetTest(RowSet*, u8 iBatch, i64);
+int sqlite3RowSetTest(RowSet*, int iBatch, i64);
 int sqlite3RowSetNext(RowSet*, i64*);
 
 void sqlite3CreateView(Parse*,Token*,Token*,Token*,Select*,int,int);
